+++ VpnHood.Server.App.Net/ServerApp.cs
@@ -2,7 +2,6 @@
 using System.IO;
 using System.Runtime.InteropServices;
 using System.Threading;
-using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
 using Microsoft.Extensions.Logging;
 using NLog;
@@ -223,7 +222,9 @@ public class ServerApp : AppBaseNet<ServerApp>
                 Tracker = _googleAnalytics,
                 SystemInfoProvider = systemInfoProvider,
                 SocketFactory = new ServerSocketFactory(),
-                StoragePath = InternalStoragePath
+                StoragePath = InternalStoragePath,
+                TcpConnectTimeout = AppSettings.TcpConnectTimeout,
+                MaxTcpConnectWaitCount= AppSettings.MaxTcpConnectWaitCount
             });
 
             // track