+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v2.6.327
+### Server
+* Fix: Error on centos
+* Feature: Report CPU usage to access server
+* Feature: Add TcpConnectWait control
+* Feature: Add TcpChannelCount control
+
 # v2.6.326
 ### Client
 * Feature: Windows: Compile as Win-x64. NET runtime is not required anymore.